feat(sponsorship): add auth guard and centralize access control UI
- Restrict sponsorship pages to authenticated users only
- Create reusable Forbidden template component
- Replace scattered NotFound components across app with centralized template
- Add mount check to prevent content flash before auth verification
