+++ src/components/templates/index.tsx
@@ -1,6 +1,7 @@
 import { Container } from "./Container";
-import { Modal } from "./Modal";
 import { Page } from "./Page";
+import { Forbidden } from "./Forbidden";
 import { Portal } from "./Portal";
+import { Modal } from "./Modal";
 
-export const Template = { Container, Page, Portal, Modal };
+export const Template = { Container, Page, Forbidden, Portal, Modal };
